#2997-include only notification service
