+++ src/backend/src/prisma/schema.prisma
@@ -180,9 +180,9 @@ model User {
   deletedFrequentlyAskedQuestions FrequentlyAskedQuestion[] @relation(name: "frequentlyAskedQuestionDeleter")
   createdMilestones               Milestone[]               @relation(name: "milestoneCreator")
   deletedMilestones               Milestone[]               @relation(name: "milestoneDeleter")
-  receivedAnnouncements           Announcement[]            @relation(name: "ReceivedAnnouncements")
-  createdAnnouncements            Announcement[]            @relation(name: "CreatedAnnouncements")
-  unreadNotifications             Notification[]            @relation(name: "UserNotifications")
+  receivedAnnouncements           Announcement[]            @relation(name: "receivedAnnouncements")
+  createdAnnouncements            Announcement[]            @relation(name: "createdAnnouncements")
+  unreadNotifications             Notification[]            @relation(name: "userNotifications")
 }
 
 model Role {
@@ -934,15 +934,15 @@ model Milestone {
 model Announcement {
   announcementId String   @id @default(uuid())
   text           String
-  usersReceived  User[]   @relation("ReceivedAnnouncements")
+  usersReceived  User[]   @relation("receivedAnnouncements")
   dateCrated     DateTime
   userCreatedId  String
-  userCreated    User     @relation("CreatedAnnouncements", fields: [userCreatedId], references: [userId])
+  userCreated    User     @relation("createdAnnouncements", fields: [userCreatedId], references: [userId])
 }
 
 model Notification {
   notificationId String @id @default(uuid())
   text           String
   iconName       String
-  users          User[] @relation("UserNotifications")
+  users          User[] @relation("userNotifications")
 }
