+++ src/backend/src/services/users.services.ts
@@ -38,8 +38,6 @@ import { getAuthUserQueryArgs } from '../prisma-query-args/auth-user.query-args'
 import authenticatedUserTransformer from '../transformers/auth-user.transformer';
 import { getTaskQueryArgs } from '../prisma-query-args/tasks.query-args';
 import taskTransformer from '../transformers/tasks.transformer';
-import notificationTransformer from '../transformers/notifications.transformer';
-import { getNotificationQueryArgs } from '../prisma-query-args/notifications.query-args';
 
 export default class UsersService {
   /**
@@ -567,45 +565,5 @@ export default class UsersService {
 
     const resolvedTasks = await Promise.all(tasksPromises);
     return resolvedTasks.flat();
-  }
-
-  /**
-   * Creates and sends a notification to all users with the given userIds
-   * @param text writing in the notification
-   * @param iconName icon that appears in the notification
-   * @param userIds ids of users to send the notification to
-   * @param organization
-   * @returns the created notification
-   */
-  static async sendNotifcationToUsers(text: string, iconName: string, userIds: string[], organization: Organization) {
-    const createdNotification = await prisma.notification.create({
-      data: {
-        text,
-        iconName
-      },
-      ...getNotificationQueryArgs(organization.organizationId)
-    });
-
-    if (!createdNotification) throw new HttpException(500, 'Failed to create notification');
-
-    const notificationsPromises = userIds.map(async (userId) => {
-      const requestedUser = await prisma.user.findUnique({
-        where: { userId }
-      });
-
-      if (!requestedUser) throw new NotFoundException('User', userId);
-
-      return await prisma.user.update({
-        where: { userId: requestedUser.userId },
-        data: {
-          unreadNotifications: {
-            connect: { notificationId: createdNotification.notificationId }
-          }
-        }
-      });
-    });
-
-    await Promise.all(notificationsPromises);
-    return notificationTransformer(createdNotification);
   }
 }
