+++ src/backend/tests/unmocked/users.test.ts
@@ -1,9 +1,8 @@
 import { Organization } from '@prisma/client';
 import { createTestOrganization, createTestTask, createTestUser, resetUsers } from '../test-utils';
-import { batmanAppAdmin, supermanAdmin } from '../test-data/users.test-data';
+import { batmanAppAdmin } from '../test-data/users.test-data';
 import UsersService from '../../src/services/users.services';
 import { NotFoundException } from '../../src/utils/errors.utils';
-import prisma from '../../src/prisma/prisma';
 
 describe('User Tests', () => {
   let orgId: string;
@@ -47,40 +46,6 @@ describe('User Tests', () => {
       const userTasks = await UsersService.getManyUserTasks([testBatman.userId, testBatman.userId], organization);
 
       expect(userTasks).toStrictEqual([batmanTask, batmanTask]);
-    });
-  });
-
-  describe('Send Notification', () => {
-    it('fails on invalid user id', async () => {
-      await expect(
-        async () => await UsersService.sendNotifcationToUsers('test notification', 'star', ['1', '2'], organization)
-      ).rejects.toThrow(new NotFoundException('User', '1'));
-    });
-
-    it('Succeeds and sends notification to user', async () => {
-      const testBatman = await createTestUser(batmanAppAdmin, orgId);
-      const testSuperman = await createTestUser(supermanAdmin, orgId);
-      await UsersService.sendNotifcationToUsers(
-        'test notification',
-        'star',
-        [testBatman.userId, testSuperman.userId],
-        organization
-      );
-
-      const batmanWithNotifications = await prisma.user.findUnique({
-        where: { userId: testBatman.userId },
-        include: { unreadNotifications: true }
-      });
-
-      const supermanWithNotifications = await prisma.user.findUnique({
-        where: { userId: testBatman.userId },
-        include: { unreadNotifications: true }
-      });
-
-      expect(batmanWithNotifications?.unreadNotifications).toHaveLength(1);
-      expect(batmanWithNotifications?.unreadNotifications[0].text).toBe('test notification');
-      expect(supermanWithNotifications?.unreadNotifications).toHaveLength(1);
-      expect(supermanWithNotifications?.unreadNotifications[0].text).toBe('test notification');
     });
   });
 });
